wip (#1352): added voting handler and repair format
